# docs/openapi/ai.json
{
  "components": {
    "schemas": {
...
      "CommitInput": {
        "additionalProperties": false,
        "description": "Spring이 GitHub에서 수집해 전달하는 정규화 커밋.",
        "properties": {
          "additions": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "그룹화 단계에서는 diff를 조회하지 않으므로 비어 있을 수 있다",
            "title": "Additions"
          },
          "author_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Email"
          },
          "author_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Login"
          },
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "GitHub 계정이 연결되지 않은 커밋의 작성자 이름",
            "title": "Author Name"
          },
          "authored_at": {
            "format": "date-time",
            "title": "Authored At",
            "type": "string"
          },
          "changed_files": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "변경 파일 수. files가 비어 있을 때 점수에 사용한다",
            "title": "Changed Files"
          },
          "deletions": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "그룹화 단계에서는 diff를 조회하지 않으므로 비어 있을 수 있다",
            "title": "Deletions"
          },
          "exclusion_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExclusionReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ChangedFile"
            },
            "title": "Files",
            "type": "array"
          },
          "is_excluded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "수집기가 판정한 제외 여부. null이면 셀렉터가 규칙으로 판정한다",
            "title": "Is Excluded"
          },
          "issue_numbers": {
            "items": {
              "exclusiveMinimum": 0.0,
              "type": "integer"
            },
            "title": "Issue Numbers",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "parent_count": {
            "minimum": 0.0,
            "title": "Parent Count",
            "type": "integer"
          },
          "pull_request_number": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pull Request Number"
          },
          "repository_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id"
          },
          "sha": {
            "maxLength": 40,
            "minLength": 7,
            "title": "Sha",
            "type": "string"
          }
        },
        "required": [
          "sha",
          "message",
          "authored_at",
          "parent_count"
        ],
        "title": "CommitInput",
        "type": "object"
      },
...
      "DependencyFile": {
        "additionalProperties": false,
        "description": "기술 스택 판별에 필요한 의존성·설정 파일.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "DependencyFile",
        "type": "object"
      },
...
      "DiffEvidenceRequest": {
        "additionalProperties": false,
        "description": "선택된 후보의 diff를 조회해 커밋별 근거로 요약하도록 요청한다.\n\n후보 지정 방식은 ``/internal/collect/candidate-details``와 같고, 경험 제목만 더한다.",
        "properties": {
          "commit_shas": {
            "description": "선택된 후보에 포함된 commit SHA. PR 전체면 빈 배열 가능",
            "items": {
              "maxLength": 40,
              "minLength": 7,
              "pattern": "^[0-9a-fA-F]+$",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Commit Shas",
            "type": "array"
          },
          "github_pr_number": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "PR 기반 후보의 GitHub PR 번호",
            "title": "Github Pr Number"
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryTarget"
          },
          "title": {
            "description": "경험 후보 제목. 모델이 커밋을 어떤 경험의 일부로 읽을지 정하는 문맥",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "user_repository_id": {
            "exclusiveMinimum": 0.0,
            "title": "User Repository Id",
            "type": "integer"
          }
        },
        "required": [
          "user_repository_id",
          "repository",
          "title"
        ],
        "title": "DiffEvidenceRequest",
        "type": "object"
      },
      "DiffEvidenceResult": {
        "description": "후보의 커밋별 diff 근거. 그대로 ``/internal/analysis/star``의 evidence로 넘긴다.",
        "properties": {
          "evidence": {
            "description": "입력 커밋 순서(PR이면 PR 커밋 순서)의 커밋별 근거",
            "items": {
              "$ref": "#/components/schemas/DiffEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "github_pr_number": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Pr Number"
          },
          "partial": {
            "default": false,
            "description": "diff 수집이 상한·rate limit으로 일부만 됐는지. 요약은 받은 범위만 다룬다",
            "title": "Partial",
            "type": "boolean"
          },
          "partial_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_repository_id": {
            "exclusiveMinimum": 0.0,
            "title": "User Repository Id",
            "type": "integer"
          }
        },
        "required": [
          "user_repository_id"
        ],
        "title": "DiffEvidenceResult",
        "type": "object"
      },
...
      "Envelope_DiffEvidenceResult_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiffEvidenceResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[DiffEvidenceResult]",
        "type": "object"
      },
      "Envelope_ExperienceGroupingResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperienceGroupingResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[ExperienceGroupingResponse]",
        "type": "object"
      },
...
      "Envelope_StarAnalysisResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StarAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[StarAnalysisResponse]",
        "type": "object"
      },
      "ErrorDetail": {
        "description": "공통 에러 형식(0-3)의 error 필드.",
        "properties": {
          "code": {
            "description": "에러 코드(0-4 표 기준)",
            "enum": [
              "INVALID_PAYLOAD",
              "UNAUTHORIZED",
              "RESOURCE_NOT_FOUND",
              "NO_TOOL_CALL",
              "EVIDENCE_MISMATCH",
              "LLM_RATE_LIMIT",
              "GITHUB_API_ERROR",
              "LLM_UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "사람이 읽을 에러 메시지",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "description": "재시도로 해결 가능한지 여부",
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "실패 응답의 실제 필드 관계를 OpenAPI에 표현한다.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "const": false,
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "meta",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ExclusionReason": {
        "description": "본인 기여 후보에서 제외한 결정적 이유.",
        "enum": [
          "BOT",
          "MERGE_COMMIT",
          "LOCKFILE_ONLY",
          "NOT_OWN"
        ],
        "title": "ExclusionReason",
        "type": "string"
      },
...
      "ExperienceGroupingRequest": {
        "additionalProperties": false,
        "description": "커밋 선별과 경험 그룹화를 요청한다.",
        "properties": {
          "collection_partial": {
            "default": false,
            "description": "수집 결과(RepositoryCollectionResult.partial)가 일부만 반환됐는지. true면 후보가 있어도 verdict는 PARTIAL",
            "title": "Collection Partial",
            "type": "boolean"
          },
          "commits": {
            "items": {
              "$ref": "#/components/schemas/CommitInput"
            },
            "title": "Commits",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/schemas__analysis__IssueContext"
            },
            "title": "Issues",
            "type": "array"
          },
          "pull_requests": {
            "items": {
              "$ref": "#/components/schemas/PullRequestContext"
            },
            "title": "Pull Requests",
            "type": "array"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "target_login": {
            "minLength": 1,
            "title": "Target Login",
            "type": "string"
          }
        },
        "required": [
          "repository_id",
          "target_login",
          "commits"
        ],
        "title": "ExperienceGroupingRequest",
        "type": "object"
      },
      "ExperienceGroupingResponse": {
        "description": "사용자가 확정할 수 있는 경험 후보 목록.",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ExperienceCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "excluded_commit_shas": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Commit Shas",
            "type": "array"
          },
          "verdict": {
            "enum": [
              "OK",
              "EMPTY",
              "PARTIAL"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict"
        ],
        "title": "ExperienceGroupingResponse",
        "type": "object"
      },
...
      "Meta": {
        "description": "요청 처리 메타데이터(로깅/추적용).",
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "실제 사용된 모델(LLM 미사용 응답은 null 또는 처리 방식 이름)",
            "title": "Model"
          },
          "processing_ms": {
            "default": 0,
            "description": "처리 소요 시간(ms)",
            "title": "Processing Ms",
            "type": "integer"
          },
          "tool_calls_made": {
            "default": 0,
            "description": "이번 처리에서 실제 도구 호출 횟수",
            "title": "Tool Calls Made",
            "type": "integer"
          }
        },
        "title": "Meta",
        "type": "object"
      },
...
      "PartialReason": {
        "description": "수집 결과가 일부만 반환된 이유.",
        "enum": [
          "GITHUB_RATE_LIMITED",
          "CAP_EXCEEDED"
        ],
        "title": "PartialReason",
        "type": "string"
      },
      "PullRequestContext": {
        "additionalProperties": false,
        "description": "PR 후보의 제목·설명과 Issue 연결에 쓰는 PR 문맥(Spring ``pull_request`` 행).",
        "properties": {
          "body_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Excerpt"
          },
          "linked_issue_numbers": {
            "description": "PR 본문의 Resolves #n 등으로 연결된 Issue",
            "items": {
              "exclusiveMinimum": 0.0,
              "type": "integer"
            },
            "title": "Linked Issue Numbers",
            "type": "array"
          },
          "number": {
            "exclusiveMinimum": 0.0,
            "title": "Number",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "number",
          "title"
        ],
        "title": "PullRequestContext",
        "type": "object"
      },
...
      "RepositoryTarget": {
        "additionalProperties": false,
        "description": "Spring이 소유권 검증을 마친 수집 대상 저장소.",
        "properties": {
          "default_branch": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Default Branch",
            "type": "string"
          },
          "github_repo_id": {
            "exclusiveMinimum": 0.0,
            "title": "Github Repo Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_login": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Owner Login",
            "type": "string"
          }
        },
        "required": [
          "github_repo_id",
          "owner_login",
          "name",
          "default_branch"
        ],
        "title": "RepositoryTarget",
        "type": "object"
      },
...
      "StarAnalysisRequest": {
        "additionalProperties": false,
        "description": "확정된 경험의 STAR 분석을 요청한다.",
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "confirmed_answers": {
            "items": {
              "type": "string"
            },
            "title": "Confirmed Answers",
            "type": "array"
          },
          "dependency_files": {
            "items": {
              "$ref": "#/components/schemas/DependencyFile"
            },
            "title": "Dependency Files",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/DiffEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "source_type": {
            "enum": [
              "PR",
              "ISSUE",
              "COMMIT_CLUSTER",
              "MANUAL"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "target_login": {
            "title": "Target Login",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "target_login",
          "title",
          "source_type",
          "evidence"
        ],
        "title": "StarAnalysisRequest",
        "type": "object"
      },
...
      "schemas__analysis__IssueContext": {
        "additionalProperties": false,
        "description": "Issue 후보의 제목·설명에 쓰는 Issue 문맥(Spring ``issue`` 행).",
        "properties": {
          "body_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Excerpt"
          },
          "number": {
            "exclusiveMinimum": 0.0,
            "title": "Number",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "number",
          "title"
        ],
        "title": "IssueContext",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI는 판단·생성만 수행하며 인증·소유권 검증·영속 저장은 Spring이 담당합니다. `/internal/*` 응답은 공통 `Envelope`이고 스키마 검증 실패는 422 대신 400입니다.\n\n**현재 구현 범위:** GitHub 수집과 템플릿/규칙 기반 B-1/B-2가 구현되어 있습니다. A 경험 그룹화·STAR 생성은 아직 미구현이며, 기본 파이프라인 호출은 500으로 끝납니다. `/health`는 프로세스 생존 확인이며 LLM 또는 A 기능 readiness 검사가 아닙니다. 이 API는 내부 서버 간 호출용이고 인터넷에 공개하는 인증 API가 아닙니다.",
    "summary": "Spring 백엔드에서 호출하는 stateless 수집·분석·되묻기 API",
    "title": "Gitory AI Internal API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/analysis/diff-evidence": {
      "post": {
        "description": "선택된 후보의 diff를 GitHub에서 다시 조회해 커밋별 변경 요약·기술 포인트로 정리합니다. 응답의 evidence는 그대로 /internal/analysis/star 요청의 evidence로 넘깁니다. 원본 patch는 AI 안에서만 쓰고 응답·로그·DB에 남기지 않습니다. 비밀값은 모델로 보내기 전에 가립니다. diff 수집이 상한이나 rate limit으로 일부만 되면 200의 partial=true로 표시합니다. GitHub 토큰은 헤더로만 전달합니다.",
        "operationId": "collect_diff_evidence_internal_analysis_diff_evidence_post",
        "parameters": [
          {
            "description": "Spring이 호출 중에만 전달하는 GitHub OAuth 토큰. 요청 본문이나 문서 예시에 넣지 않습니다.",
            "in": "header",
            "name": "X-GitHub-Token",
            "required": true,
            "schema": {
              "description": "Spring이 호출 중에만 전달하는 GitHub OAuth 토큰. 요청 본문이나 문서 예시에 넣지 않습니다.",
              "minLength": 1,
              "title": "X-Github-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DiffEvidenceResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_PAYLOAD",
                    "message": "요청 스키마가 명세와 일치하지 않습니다.",
                    "retryable": false
                  },
                  "meta": {
                    "processing_ms": 0,
                    "tool_calls_made": 0
                  },
                  "success": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "INVALID_PAYLOAD: 요청 본문 또는 필수 헤더 검증 실패. 재시도 불가."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND: GitHub 리소스가 없거나 토큰 권한으로 읽을 수 없음."
          },
          "500": {
            "content": {
              "text/plain": {
                "example": "Internal Server Error",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "분석 처리 중 내부 오류 또는 LLM 호출 실패"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "GITHUB_API_ERROR: GitHub API 호출 실패. 재시도 가능."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "LLM_UNAVAILABLE: LLM 호출·출력 검증 실패(retryable=true) 또는 LLM 환경 설정 누락(retryable=false)."
          }
        },
        "summary": "A 선택 후보의 커밋별 diff 근거 생성",
        "tags": [
          "analysis"
        ]
      }
    },
    "/internal/analysis/groups": {
      "post": {
        "description": "수집된 커밋을 선별하고 PR·Issue 연결 및 LLM 분석으로 경험 후보를 묶습니다.",
        "operationId": "group_experiences_internal_analysis_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperienceGroupingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ExperienceGroupingResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_PAYLOAD",
                    "message": "요청 스키마가 명세와 일치하지 않습니다.",
                    "retryable": false
                  },
                  "meta": {
                    "processing_ms": 0,
                    "tool_calls_made": 0
                  },
                  "success": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "INVALID_PAYLOAD: 요청 본문 또는 필수 헤더 검증 실패. 재시도 불가."
          },
          "500": {
            "content": {
              "text/plain": {
                "example": "Internal Server Error",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "분석 처리 중 내부 오류 또는 LLM 호출 실패"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "LLM_UNAVAILABLE: 그룹화 LLM 호출·출력 검증 실패(retryable=true) 또는 LLM 설정 누락(retryable=false)."
          }
        },
        "summary": "A 경험 후보 그룹화",
        "tags": [
          "analysis"
        ]
      }
    },
    "/internal/analysis/star": {
      "post": {
        "description": "/internal/analysis/diff-evidence가 만든 커밋별 근거로 인터뷰 전 STAR 초안을 생성합니다. 근거 커밋이 없거나 입력에 없는 수치를 쓴 문장은 비우고 removed_claims에 남깁니다. 비운 칸은 missing_fields로 반환하며 B 인터뷰가 이 칸을 질문합니다.",
        "operationId": "analyze_star_internal_analysis_star_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StarAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_StarAnalysisResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_PAYLOAD",
                    "message": "요청 스키마가 명세와 일치하지 않습니다.",
                    "retryable": false
                  },
                  "meta": {
                    "processing_ms": 0,
                    "tool_calls_made": 0
                  },
                  "success": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "INVALID_PAYLOAD: 요청 본문 또는 필수 헤더 검증 실패. 재시도 불가."
          },
          "500": {
            "content": {
              "text/plain": {
                "example": "Internal Server Error",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "분석 처리 중 내부 오류 또는 LLM 호출 실패"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "LLM_UNAVAILABLE: LLM 호출·출력 검증 실패(retryable=true) 또는 LLM 환경 설정 누락(retryable=false)."
          }
        },
        "summary": "A 확정 경험의 STAR 초안 생성",
        "tags": [
          "analysis"
        ]
      }
    },
...
  },
...
}